% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AMECO_spring_2023}
\alias{AMECO_spring_2023}
\title{AMECO_spring_2023}
\format{
A data frame (33407 obs of 70 variables):
}
\usage{
AMECO_spring_2023
}
\description{
Spring 2023 Economic Forecast (updated 15/05/2023).
}
\keyword{datasets}
