% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AMECO_spring_2024}
\alias{AMECO_spring_2024}
\title{AMECO_spring_2024}
\format{
A data frame (32637 obs of 71 variables):
\describe{
\item{CODE}{Code}
\item{COUNTRY}{Country}}
\item{SUB_CHAPTER}{Sub-chapter}}
\item{TITLE}{Title}}
\item{UNIT}{Unit}}
}
\usage{
AMECO_spring_2024
}
\description{
Spring 2024 Economic Forecast (updated 15/05/2024).
}
\keyword{datasets}
