% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_series.R
\name{find_series}
\alias{find_series}
\title{Find a series in the AMECO data}
\usage{
find_series(data, text, field)
}
\arguments{
\item{data}{A data.frame}

\item{text}{A string}

\item{field}{An integer (1-4)}
}
\value{
A data.frame
}
\description{
Find a series in the AMECO data
}
\examples{
ZUTN <- find_series(AMECO_spring_2024, "zutn", field = 1)
}
