% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ecmacro.R
\name{get_ecmacro}
\alias{get_ecmacro}
\title{Download the latest AMECO data}
\usage{
get_ecmacro(url)
}
\arguments{
\item{url}{A string}
}
\value{
A data.frame
}
\description{
Download the latest AMECO data
}
\examples{
AMECO <- get_ecmacro()
}
