% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series.R
\name{get_series}
\alias{get_series}
\title{Get a series from the AMECO data}
\usage{
get_series(ameco, code)
}
\arguments{
\item{ameco}{A data.frame}

\item{code}{A string}
}
\value{
A ts object
}
\description{
Get a series from the AMECO data
}
